Move Tag utility methods into Tag class
Moved static methods tagSetToString and stringToTagSet from TagCommand
and TagCommandParser into Tag class for cohesiveness
