+++ src/main/java/seedu/address/logic/commands/ListTagsCommand.java
@@ -1,0 +1,4 @@
+package seedu.address.logic.commands;
+
+public class ListTagsCommand {
+}
