+++ src/main/java/seedu/address/logic/commands/TagCommand.java
@@ -3,7 +3,6 @@ package seedu.address.logic.commands;
 import static java.util.Objects.requireNonNull;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 
-import java.util.Iterator;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
@@ -54,7 +53,7 @@ public class TagCommand extends Command {
         }
 
         Person personToTag = lastShownList.get(targetIndex.getZeroBased());
-        String addedTagsString = tagSetToString(addedTags);
+        String addedTagsString = Tag.tagSetToString(addedTags);
         // Union of existing tags and new tags
         addedTags.addAll(personToTag.getTags());
 
@@ -67,23 +66,6 @@ public class TagCommand extends Command {
         return new CommandResult(String.format(MESSAGE_TAG_PERSON_SUCCESS, personToTag.getName(), addedTagsString));
     }
 
-    /**
-     * Utility method to convert Set of Tags to strings delimited by commas
-     * @param tagList The Set of Tags to convert
-     * @return String of tags delimited by commas
-     */
-    public static String tagSetToString(Set<Tag> tagList) {
-        StringBuilder stringBuilder = new StringBuilder();
-        Iterator<Tag> it = tagList.iterator();
-        for (int i = 0; i < tagList.size(); i++) {
-            Tag t = it.next();
-            stringBuilder.append(t);
-            if (i != tagList.size() - 1) {
-                stringBuilder.append(", ");
-            }
-        }
-        return stringBuilder.toString();
-    }
 
     @Override
     public boolean equals(Object other) {
